Added Operations team charter.
This is part of the DSF's effort to standarize the technical teams.

The Ops team didn't have much in the way of written governance in the
past. The expectation is that this will help future team members get
onboarded. The expectations will also help the DSF Board and other
groups support the team.